requirements.txt
flask
requests
chromadb
pdfminer.six
celery
SQLAlchemy
chatwrap @ git+https://github.com/stefanbertici/chatwrap.git